refactor: extract inline components
